+++ CHANGELOG.md
@@ -11,6 +11,7 @@ v0.15.19 (??.??.????)
  * new Wrapper function tx_rnbase_parameters::getPostAndGetParametersMerged for _GPmerged
  * new Wrapper function tx_rnbase_parameters::getPostOrGetParameter for _GP
  * new Wrapper function tx_rnbase_util_Link::linkThisScript
+ * new Wrapper function tx_rnbase_util_Files::writeFile
 
 v0.15.18 (02.12.2015)
  * Bugfix: tx_rnbase_util_Dates::date_mysql2tstamp liefert nun immer einen Integer und nicht NULL im Fehlerfall